+++ CHANGES.rst
@@ -31,6 +31,7 @@ Improvements
 - Automatically paste obvious email addresses into the email field when pasting in
   the user search dialog (:pr:`7538`)
 - Allow changing/removing the registration fee of pending registrations (:pr:`7572`)
+- Add QR code generator to event share widget (:issue:`6796`, :pr:`7504`)
 
 Bugfixes
 ^^^^^^^^